+++ src/Chat.Host/Chat.Host.csproj
@@ -18,8 +18,8 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="tanka.graphql" Version="0.5.1-beta.22" />
-    <PackageReference Include="tanka.graphql.server" Version="0.5.1-beta.22" />
+    <PackageReference Include="tanka.graphql" Version="0.5.1-beta.23" />
+    <PackageReference Include="tanka.graphql.server" Version="0.5.1-beta.23" />
   </ItemGroup>
 
 </Project>
