+++ src/Host/ClientApp/package.json
@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^3.9.2",
+    "@material-ui/core": "^3.9.3",
     "@material-ui/icons": "^3.0.2",
     "@tanka/tanka-graphql-server-link": "^0.5.0",
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
     "apollo-link": "^1.2.11",
     "apollo-link-error": "^1.1.10",
-    "graphql": "^14.1.1",
+    "graphql": "^14.2.1",
     "graphql-tag": "^2.10.1",
-    "react": "^16.8.5",
+    "react": "^16.8.6",
     "react-apollo": "^2.5.3",
-    "react-dom": "^16.8.5",
-    "react-router-bootstrap": "^0.24.4",
+    "react-dom": "^16.8.6",
+    "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "^5.0.0",
     "react-scripts": "^2.1.8",
     "rimraf": "^2.6.3"
