+++ src/Host/Host.csproj
@@ -17,8 +17,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="1.1.0" />
-    <PackageReference Include="tanka.graphql" Version="0.5.1-beta.22" />
-    <PackageReference Include="tanka.graphql.server" Version="0.5.1-beta.22" />
+    <PackageReference Include="tanka.graphql" Version="0.5.1-beta.23" />
+    <PackageReference Include="tanka.graphql.server" Version="0.5.1-beta.23" />
   </ItemGroup>
 
   <ItemGroup>
